删除角色前判断是否有用户

+++ model/AuthRole.php
@@ -21,9 +21,18 @@ class AuthRole extends \think\Model
 
     /**
      * 关联删除 AuthAccess
+     * 判断是否有用户使用此角色,如果有返回使用角色数量
+     * 否则删除角色数据,调用authAccess方法如果有数据删除关联AuthAccess模型数据
+     *
      * @return bool
      */
-    public function authRoleDelete(){
+    public function authRoleDelete()
+    {
+        $roleCount = AuthRoleUser::where(['role_id'=>$this->id])->count();
+        if($roleCount > 0){
+            return "已有{$roleCount}用户在是有此角色不可删除";
+        }
+
         if($this->delete()){
             if($this->authAccess){
                 AuthAccess::where(['role_id'=>$this->id,'type'=>'admin_url'])->delete();
